backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
websockets==13.1
faster-whisper==1.0.3
sounddevice==0.5.0
numpy>=1.26.4
TTS==0.22.0
pywebview==5.3
pystray==0.19.5
pillow==10.4.0
google-generativeai==0.8.3
requests==2.32.3
python-dotenv==1.0.1
aiofiles==24.1.0
scipy>=1.13.1
pvporcupine==3.0.0
openai>=1.0.0
